Revert "perf(voxel_grid_downsample_filter): performance tuning (#3679)"
This reverts commit edcad42338a56cd2c7eaf0803f0861c56493608c.
