+++ sensing/pointcloud_preprocessor/include/pointcloud_preprocessor/downsample_filter/voxel_grid_downsample_filter_nodelet.hpp
@@ -1,4 +1,4 @@
-// Copyright 2023 TIER IV, Inc.
+// Copyright 2020 Tier IV, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -52,7 +52,6 @@
 #define POINTCLOUD_PREPROCESSOR__DOWNSAMPLE_FILTER__VOXEL_GRID_DOWNSAMPLE_FILTER_NODELET_HPP_
 
 #include "pointcloud_preprocessor/filter.hpp"
-#include "pointcloud_preprocessor/transform_info.hpp"
 
 #include <pcl/filters/voxel_grid.h>
 #include <pcl/search/pcl_search.h>
@@ -67,16 +66,10 @@ protected:
   void filter(
     const PointCloud2ConstPtr & input, const IndicesPtr & indices, PointCloud2 & output) override;
 
-  // TODO(atsushi421): Temporary Implementation: Remove this interface when all the filter nodes
-  // conform to new API
-  virtual void faster_filter(
-    const PointCloud2ConstPtr & input, const IndicesPtr & indices, PointCloud2 & output,
-    const TransformInfo & transform_info);
-
 private:
-  float voxel_size_x_;
-  float voxel_size_y_;
-  float voxel_size_z_;
+  double voxel_size_x_;
+  double voxel_size_y_;
+  double voxel_size_z_;
 
   /** \brief Parameter service callback result : needed to be hold */
   OnSetParametersCallbackHandle::SharedPtr set_param_res_;
