+++ sensing/pointcloud_preprocessor/include/pointcloud_preprocessor/filter.hpp
@@ -52,8 +52,6 @@
 #ifndef POINTCLOUD_PREPROCESSOR__FILTER_HPP_
 #define POINTCLOUD_PREPROCESSOR__FILTER_HPP_
 
-#include "pointcloud_preprocessor/transform_info.hpp"
-
 #include <memory>
 #include <string>
 #include <vector>
@@ -136,6 +134,18 @@ public:
     const rclcpp::NodeOptions & options = rclcpp::NodeOptions());
 
 protected:
+  struct TransformInfo
+  {
+    TransformInfo()
+    {
+      eigen_transform = Eigen::Matrix4f::Identity(4, 4);
+      need_transform = false;
+    }
+
+    Eigen::Matrix4f eigen_transform;
+    bool need_transform;
+  };
+
   /** \brief The input PointCloud2 subscriber. */
   rclcpp::Subscription<PointCloud2>::SharedPtr sub_input_;
