+++ sensing/pointcloud_preprocessor/include/pointcloud_preprocessor/outlier_filter/ring_outlier_filter_nodelet.hpp
@@ -17,7 +17,6 @@
 
 #include "autoware_point_types/types.hpp"
 #include "pointcloud_preprocessor/filter.hpp"
-#include "pointcloud_preprocessor/transform_info.hpp"
 
 #include <point_cloud_msg_wrapper/point_cloud_msg_wrapper.hpp>
